New Fawkes LiveCD & Package/Service Role
Important:

- [fawkes] MTL-2198 New Fawkes LiveCD
- [ansible] New `packages` and `services` role
- [fawkes] Disabled ipv6
- [fawkes] New crucible 0.0.5
- [ansible] New bootserver role for setting up the management-vm's services

General:

- Ansible Lint action change; ansible-community/ansible-lint-action is
  no longer maintained. Using ansible/ansible-lint instead, along with
  newer version.
- (MTL-2197) Added a function for creating a release file
- (MTL-2193) Fixed `[FAILED] Failed to start Remount Root and Kernel File Systems.` on the LiveCD. `systemd-remount-fs.service` was unhappy it could not find `ROOTRAID` but it should be looking for `cow`.
- (MTL-2193) Fixed `[FAILED] Failed to start Rotate log files.` on all boxes, even though I only saw this on the LiveCD and Management-VM. `logrotate.service` was unhappy that some directories it expected from configurations in `/etc/logrotate.d/` did not exist. A new `fix_logrotate_errors` function is invoked in `cleanup.sh`.
- (MTL-2193) Fixed `[FAILED] Failed to start Apply Kernel Variables.` on all nodes. `systemd-sysctl.service` was failing to start due to two invalid settings in `99-tcp.conf` for `net.ipv4.tcp_rmem` and `net.ipv4.tcp_wmem` which had incorrectly formatted values.
- Shrinks base image, shrinks the following items:
  - `/usr/lib/locale` from `215M` down to `4.4M` by removing `glibc-locale` and keeping only `glibc-locale-base`
- Removal of `patterns=base` from `autoinst.template.xml`, no longer forcing the installation of several packages from the base pattern. This has some strange behavior, as "base" will still report as "Installed" but with a significant subset of packages.

New Package Handling:

- Packages and services are now defined in `/vars` and `/group_vars` and are handled by the new "packages" and "services" roles
- Removed all `role/*/vars/{packages,services}` files and their associated tasks, thus removing all duplication of package/service handling
- Packages key off of inventory via `/group_vars`, or they be included in the base runbook through `/vars`
- Packages for Google and Metal are pulled out of common again, defined by `/group_vars/google` and `/group_vars/metal`

+++ scripts/packer/cleanup-functions.sh
@@ -121,7 +121,7 @@ function cleanup_package_manager {
     esac
 }
 
-# Function for removing udev and networkmanager files from the build environment.
+# Function for removing udev and network manager files from the build environment.
 function cleanup_network {
     echo 'Removing build environment network files ... '
     echo '- Purge persistent-net udev rules'
@@ -134,8 +134,36 @@ function cleanup_network {
             :
             ;;
         suse)
-            echo '- Purge wicked interface cache (necessary for virsh, or `vagrant up` on multiple instances will fail)'
-            rm -f /var/lib/wicked/*.xml
+            if [ -d /var/lib/wicked ]; then
+                echo '- Purge wicked interface cache (necessary for virsh, or `vagrant up` on multiple instances will fail)'
+                rm -f /var/lib/wicked/*.xml
+            fi
+            ;;
+        *)
+            echo >&2 'Unhandled OS; nothing to do'
+            ;;
+    esac
+    echo 'Done'
+}
+
+
+# Function for removing network files on metal media.
+function cleanup_metal_network {
+    echo 'Removing build environment network files for metal ... '
+    case "${OS}" in
+        debian)
+            :
+            ;;
+        rhel)
+            :
+            ;;
+        suse)
+            echo '- Purging "eth" sysconfig files'
+            rm -f /etc/sysconfig/network/*eth*
+            if [ -d /etc/NetworkManager/system-connections ]; then
+                echo '- Purging NetworkManager "eth" files'
+                rm -f /etc/NetworkManager/system-connections/*eth*
+            fi
             ;;
         *)
             echo >&2 'Unhandled OS; nothing to do'
@@ -183,4 +211,66 @@ function defrag {
     filler="$(($(df -BM --output=avail /|grep -v Avail|cut -d "M" -f1)-1024))"
     dd if=/dev/zero of=/root/zero-file bs=1M count=$filler
     rm -f /root/zero-file
+}
+
+# Fix ``logrotate.service``, preventing confusing failure messages 
+# during login such as ``[FAILED] Failed to start Rotate log files.``.
+function fix_logrotate_errors {
+    local logrotate=/etc/logrotate.d
+    local olddirs
+    echo "Scanning logrotate configurations for directory statements ... "
+    if [ ! -d "$logrotate" ]; then
+        echo "Nothing to do, $logrotate does not exist."
+        return
+    fi
+    if ! grep -q olddir "$logrotate/"* ; then
+        return
+    fi
+    mapfile -t olddirs < <(grep olddir $logrotate/* | awk '{print $NF}')
+    echo "Fixing logrotate.service; validating [${#olddirs[@]}] olddir line(s) ... "
+    for olddir in "${olddirs[@]}"; do
+        if [ -d "$olddir" ]; then
+            echo "Logrotate directory [$olddir] already exists. Continuing ... "
+            continue
+        else
+            echo "Logrotate directory [$olddir] does not exist! Creating ... "
+            mkdir -p "$olddir"
+        fi
+    done
+    echo 'Done.'
+}
+
+# Fix ``systemd-remount-fs.service``, preventing confusing failure messages
+# during login like ``[FAILED] Failed to start Remount Root and Kernel File Systems.``
+function fix_livecd_systemd_remount {
+    sed -i -E 's:^(LABEL=)\w+(\s+/\s+):\1cow\2:' /etc/fstab
+}
+
+function create_release_file {
+    local name
+    local artifact_version
+    local epoch
+    local hash
+    local timestamp
+    name="${1:-''}"
+    artifact_version="${2:-''}"
+    if [ -z "${name}" ]; then
+        echo >&2 'Missing name for release file. Aborting!'
+        return 1
+    fi
+    echo "Making /etc/$name-release ... "
+    if [[ -z "$artifact_version" ]] || [[ "$artifact_version" = 'none' ]]; then
+      hash="dev"
+      epoch="$(date -u +%s%N | cut -b1-13)"
+    else
+      hash="$(echo "$artifact_version" | awk -F- '{print $1}')"
+      epoch="$(echo "$artifact_version" | awk -F- '{print $NF}')"
+    fi
+    timestamp="$(date -d "@${epoch:0:-3}" '+%Y:%m:%d-%H:%M:%S')"
+    cat << EOF > "/etc/${name}-release"
+VERSION=$hash-$epoch
+TIMESTAMP=$timestamp
+EOF
+    echo 'Done. Preview:'
+    cat "/etc/${name}-release"
 }
